requirements.txt
# Deep Learning and Computer Vision
torch>=1.8.0
torchvision>=0.9.0
ultralytics>=8.0.0
opencv-python>=4.5.0

# Image Processing
Pillow>=8.0.0
numpy>=1.19.0

# Web and Networking
Flask>=2.0.0
requests>=2.23.0

# Visualization and Simulation
pygame>=2.0.0
matplotlib>=3.3.0

# Data Processing
pandas>=1.1.4
scipy>=1.4.1
scikit-learn>=0.24.0

# Development Tools
jupyter>=1.0.0
ipython>=7.0.0
tqdm>=4.64.0

# Configuration
PyYAML>=5.3.1

# Additional utilities
python-dateutil>=2.7.0
